services/post-service/requirements.txt
grpcio
googleapis-common-protos
sqlalchemy
psycopg2-binary
grpcio-tools 
protobuf
protoletariat
